requirements.txt
django==3.1
djangorestframework==3.11.0
factory-boy==2.12.0
faker==2.0.2
psycopg2-binary>=2.8